CMake Phase1
Too many changes. This is bad. Today I am a bad person.

Directory structure re-organized.
CMakeLists.txt added to root directory and source directory.
The library compiles from the build directory.

To build: make a ChordAlg/build directory, then type:
cmake -G "Unix Makefiles" ..

Unit tests not integrated yet.
Executables not integrated yet.

To please warnings, most int's became size_t's.
Another warning caught a bug in tree_represntation.cpp (unintended name clashing with variable $v$).

+++ src/elimination_order.cpp
@@ -106,7 +106,7 @@ int EliminationOrder::ComputeFill() {
     follower.resize(G_.order());
     VertexVector index;
     index.resize(G_.order());
-    for (int i = 0; i < G_.order(); ++i) {
+    for (size_t i = 0; i < G_.order(); ++i) {
         Vertex w = VertexAt(i);
         follower[w] = w;
         index[w] = i;
@@ -142,7 +142,7 @@ bool EliminationOrder::ZeroFill() const {
     follower.resize(G_.order());
     VertexVector index;
     index.resize(G_.order());
-    for (int i = 0; i < G_.order(); ++i) {
+    for (size_t i = 0; i < G_.order(); ++i) {
         Vertex w = VertexAt(i);
         follower[w] = w;
         index[w] = i;
@@ -201,7 +201,7 @@ void EliminationOrder::SetOrder(VertexVector pi) {
         std::cerr << "order does not match" << std::endl;
         exit(EXIT_FAILURE);
     }
-    for (int i = 0; i < pi.size(); ++i) {
+    for (size_t i = 0; i < pi.size(); ++i) {
         Emplace(pi[i], i);
     }
     return;
@@ -249,7 +249,7 @@ void EliminationAlgorithm::Init() {
 }
 
 void EliminationAlgorithm::Elimination() {
-    for (int i = 0; i < G_.order(); ++i) {
+    for (size_t i = 0; i < G_.order(); ++i) {
         VertexCost vc = ArgMin();
         fill_cost_ += vc.second;
         Vertex v = vc.first;
